🚚[move]: 메서드IImageView extension 파일로 분리
fatchImage 메서드 URLSessionProvider -> UIImageView Extension
