+++ OpenMarket/OpenMarket/Network/URLSessionProvider.swift
@@ -97,27 +97,3 @@ struct URLSessionProvider<T: Decodable> {
             task.resume()
         }
 }
-
-
-extension UIImageView {
-    func fetchImage(url: URL, completion: @escaping (UIImage) -> Void) {
-
-        URLSession.shared.dataTask(with: url) { data, response, _ in
-            guard let response = response as? HTTPURLResponse,
-                  (200...299).contains(response.statusCode) else {
-                return
-            }
-            
-            guard let data = data else {
-                return
-            }
-            
-            guard let image = UIImage(data: data) else {
-                return
-            }
-            
-            completion(image)
-            
-        }.resume()
-    }
-}
